Backend/Neo4j/requirements.txt
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
cryptography==2.9.2
firebase-admin==4.3.0
Flask==1.1.2
flask-restplus==0.13.0
google==2.0.3
google-api-core==1.17.0
google-api-python-client==1.8.3
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.28.1
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.29.0
httplib2==0.17.3
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
msgpack==1.0.0
neo4j==1.7.6
neobolt==1.7.17
neotime==1.7.4
oauthlib==3.1.0
prompt-toolkit==2.0.10
protobuf==3.12.0
py2neo==5.0b1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.3.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyrsistent==0.16.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
Werkzeug==0.16.1
zipp==3.1.0
